+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -30,7 +30,7 @@
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#SQL_TRACING}
  * <li>{@link is.codion.swing.framework.ui.EntityEditorPanel.Config#VALID_INDICATOR}
  * <li>{@link is.codion.swing.framework.ui.EntityEditorPanel.Config#MODIFIED_INDICATOR}
- * <li>{@link is.codion.swing.framework.ui.EntityEditorPanel.Config#DEFAULT_TEXT_FIELD_COLUMNS}
+ * <li>{@link is.codion.swing.framework.ui.EntityEditorPanel.Config#TEXT_FIELD_COLUMNS}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#MODIFIED_WARNING}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#USE_SAVE_CAPTION}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#USE_FOCUS_ACTIVATION}